#1243.  Correction to r4627

+++ trunk/src/image.cpp
@@ -651,7 +651,7 @@ namespace Exiv2 {
 
     std::string binaryToString(DataBuf& buf, size_t size, size_t start /*=0*/)
     {
-        if ( size > buf.size_ ) size = buf.size_;
+        if ( size > (size_t) buf.size_ ) size = (size_t) buf.size_;
         return binaryToString(buf.pData_,size,start);
     }
